refactor transform string dispatch
